separated logic from presentation in sale edit form.. Fixed state select component bug
